+++ packages/qfai/src/core/index.ts
@@ -8,4 +8,7 @@ export * from "./validate.js";
 export * from "./version.js";
 export * from "./validators/contracts.js";
 export * from "./validators/ids.js";
+export * from "./validators/layeredTraceability.js";
+export * from "./validators/orphanProhibition.js";
+export * from "./validators/specSplitByCapability.js";
 export * from "./validators/traceability.js";