+++ packages/qfai/src/core/validators/index.ts
@@ -1,4 +1,7 @@
 export { validateDensityHints } from "./densityHints.js";
 export { validateImportLiteEvidencePresence } from "./importLite.js";
+export { validateLayeredTraceability } from "./layeredTraceability.js";
+export { validateOrphanProhibition } from "./orphanProhibition.js";
 export { validateRequireIndexShape } from "./requireIndex.js";
+export { validateSpecSplitByCapability } from "./specSplitByCapability.js";
 export { validateStatusInSpecs } from "./statusInSpecs.js";